refactor(store): Redo store initialization. Split requests realizations from store init.
