+++ internal/app/app.go
@@ -8,7 +8,8 @@ import (
 	"sso/internal/lib/config"
 	"sso/internal/lib/migrate"
 	authService "sso/internal/services/authentication"
-	"sso/internal/store/sqlite"
+	sqlstore "sso/internal/store/sql"
+	"sso/internal/store/sql/sqlite"
 )
 
 type App struct {
@@ -22,22 +23,18 @@ func NewApp(
 ) *App {
 
 	// Инициализация хранилища
-	storage, err := sqlite.New(conf.StorePath)
-	if err != nil {
-		panic(err)
-	}
+	sqliteStore := sqlite.MustInitSqlite(conf.StorePath)
+	migrate.MustMigrate(sqliteStore)
 
-	if err := migrate.Migrate(storage); err != nil {
-		panic(err)
-	}
+	store := sqlstore.NewStore(sqliteStore)
 
 	// Инициализация auth сервиса
 	authService := authService.NewService(
 		log,
 		conf.AuthenticationService,
-		storage,
-		storage,
-		storage,
+		store.AuthenticationStore(),
+		store.AuthenticationStore(),
+		store.AuthenticationStore(),
 	)
 
 	grpcapp := grpcApp.NewGrpcApp(log, conf.GrpcConfig, authService)
